# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test api",
        "description": "An API of app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Index",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "schema": {
                            "type": "object",
                            "example": {
                                "filter[id]": null,
                                "filter[title]": null,
                                "filter[description]": null,
                                "filter[price]": null,
                                "filter[old_price]": null,
                                "filter[count]": null,
                                "filter[is_infinitely]": null,
                                "filter[raiting]": null,
                                "filter[user_id]": null,
                                "filter[category_id]": null,
                                "filter[created_at]": null,
                                "filter[updated_at]": null
                            }
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "example": "id"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        },
                        "example": "20"
                    },
                    {
                        "name": "extends",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "example": "parent,children"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "old_price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "count": {
                                                "type": "number"
                                            },
                                            "is_infinitely": {
                                                "type": "boolean"
                                            },
                                            "category_id": {
                                                "type": "number"
                                            },
                                            "images": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Кухонные приборы",
                                    "description": "Описание",
                                    "price": "100.00",
                                    "old_price": "90.00",
                                    "count": "30",
                                    "is_infinitely": "false",
                                    "category_id": "1",
                                    "images": "1,2,3"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The title field is required. (and 1 more errors)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The title field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductSchema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Игрвой ноутбук 2003"
                    },
                    "description": {
                        "type": "string",
                        "example": "john@test.com"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": "2022-06-28 06:06:17"
                    },
                    "old_price": {
                        "type": "number",
                        "format": "float",
                        "example": "2022-06-28 06:06:17"
                    },
                    "count": {
                        "type": "number",
                        "example": "7"
                    },
                    "is_infinitely": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "raiting": {
                        "type": "number",
                        "format": "float",
                        "example": "5"
                    },
                    "user_id": {
                        "type": "number",
                        "example": 1
                    },
                    "category_id": {
                        "type": "number",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-06-28 06:06:17"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-06-28 06:06:17"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}